DDCE-3557 - Upgrade to scala 2.13.x, + cleanup, + update plugins, + sbt
