+++ .gitignore
@@ -1,4 +1,3 @@
-
 logs
 project/project
 project/target
@@ -21,6 +20,4 @@ node_modules/
 npm-debug.log
 yarn-debug.log
 yarn-error.log
-
-    
 /.bsp/
