+++ build.sbt
@@ -3,14 +3,14 @@ lazy val root = (project in file("."))
   .settings(
     name := "digital-tariffs-trader-performance-tests",
     version := "0.1.0",
-    scalaVersion := "2.12.16",
+    scalaVersion := "2.13.10",
     //implicitConversions & postfixOps are Gatling recommended -language settings
     scalacOptions ++= Seq("-feature", "-language:implicitConversions", "-language:postfixOps"),
     // Enabling sbt-auto-build plugin provides DefaultBuildSettings with default `testOptions` from `sbt-settings` plugin.
-    // These testOptions are not compatible with `sbt gatling:test`. So we have to override testOptions here.
+    // These testOptions are not compatible with `sbt Gatling/test`. So we have to override testOptions here.
     Test / testOptions := Seq.empty,
-    libraryDependencies ++= Dependencies.test
+    libraryDependencies ++= Dependencies()
   )
 
-addCommandAlias("scalafmtAll", "all scalafmtSbt scalafmt test:scalafmt")
-addCommandAlias("scalastyleAll", "all scalastyle test:scalastyle")
+addCommandAlias("scalafmtAll", "all scalafmtSbt scalafmt Test/scalafmt")
+addCommandAlias("scalastyleAll", "all scalastyle Test/scalastyle")
