+++ project/plugins.sbt
@@ -1,10 +1,10 @@
-resolvers += "HMRC-open-artefacts-maven" at "https://open.artefacts.tax.service.gov.uk/maven2"
-resolvers += Resolver.url("HMRC-open-artefacts-ivy", url("https://open.artefacts.tax.service.gov.uk/ivy2"))(
+resolvers += MavenRepository("HMRC-open-artefacts-maven2", "https://open.artefacts.tax.service.gov.uk/maven2")
+resolvers += Resolver.url("HMRC-open-artefacts-ivy2", url("https://open.artefacts.tax.service.gov.uk/ivy2"))(
   Resolver.ivyStylePatterns
 )
 
-addSbtPlugin("uk.gov.hmrc"      % "sbt-auto-build"        % "3.8.0")
-addSbtPlugin("io.gatling"      %% "gatling-sbt"           % "4.2.3")
-addSbtPlugin("com.timushev.sbt" % "sbt-updates"           % "0.6.3")
-addSbtPlugin("org.scalastyle"  %% "scalastyle-sbt-plugin" % "1.0.0")
-addSbtPlugin("org.scalameta"    % "sbt-scalafmt"          % "2.4.6")
+addSbtPlugin("uk.gov.hmrc"         % "sbt-auto-build" % "3.8.0")
+addSbtPlugin("io.gatling"          % "gatling-sbt"    % "4.2.4")
+addSbtPlugin("com.timushev.sbt"    % "sbt-updates"    % "0.6.3")
+addSbtPlugin("org.scalameta"       % "sbt-scalafmt"   % "2.5.0")
+addSbtPlugin("com.beautiful-scala" % "sbt-scalastyle" % "1.5.1")
