+++ run.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-sbt gatling:test
+sbt -Dperftest.runSmokeTest=false -DrunLocal=false Gatling/test
