+++ run_format_and_deps.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-sbt clean scalafmtAll scalastyleAll compile test:compile dependencyUpdates
+sbt clean scalafmtAll scalastyleAll compile Test/compile dependencyUpdates
