+++ run_smoke_test_local.sh
@@ -2,4 +2,4 @@
 
 ./run_format_and_deps.sh
 
-sbt -Dperftest.runSmokeTest=true -DrunLocal=true gatling:test
+sbt -Dperftest.runSmokeTest=true -DrunLocal=true Gatling/test
