+++ scalastyle-config.xml
@@ -1,6 +1,6 @@
 <scalastyle>
     <name>Scalastyle standard configuration</name>
-    <check level="warning" class="org.scalastyle.file.FileTabChecker" enabled="false"/>
+    <check level="warning" class="org.scalastyle.file.FileTabChecker" enabled="true"/>
     <check level="warning" class="org.scalastyle.file.FileLengthChecker" enabled="true">
         <parameters>
             <parameter name="maxFileLength"><![CDATA[800]]></parameter>
