+++ src/test/resources/application.conf
@@ -12,13 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# The default url for every service.
-# For example
-# ServicesConfiguration.baseUrlFor("helloworld-service") == https://helloworld-service.co.uk
-baseUrl = "https://www.staging.tax.service.gov.uk"
-
 # The baseUrl can be overridden in services-local.conf. See that file for more information
-# That file is ignored when the test is not run locally. The default is true
+# That file is ignored when the test is not run locally. The default is false
 runLocal = false
 
 perftest {
@@ -45,6 +40,6 @@ perftest {
   runSmokeTest = false
 
   # Percentages of failing journeys required for a test-run to fail
-  //percentageFailureThreshold = 1
+  percentageFailureThreshold = 1
 
 }
