+++ src/test/resources/services-local.conf
@@ -12,26 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# This file is read when runLocal = true in application.conf (default)
-
-# This can be used to override the 'baseUrl'
-
-# If the service name you are searching is found, baseUrl is overridden.
-# For local development convenience, host and protocol will have the following defaults:
-#      protocol = http
-#      host = localhost
-
-# Example 1: (the base url for helloworld-service is https://127.0.0.1:9000)
-# helloworld-service {
-#      protocol = https
-#      host = 127.0.0.1
-#      port = 9000
-# }
-
-# Example 2: (the base url for helloworld-service is http://localhost:9000)
-# helloworld-service.port = 9000
-
-
 services {
   auth-login-stub.port= 9949
   binding-tariff-trader-frontend.port = 9582
