+++ src/test/scala/uk/gov/hmrc/perftests/digitaltariffs/traderui/TraderUiSimulation.scala
@@ -16,16 +16,12 @@
 
 package uk.gov.hmrc.perftests.digitaltariffs.traderui
 
-import io.gatling.http.protocol.HttpProtocolBuilder
 import uk.gov.hmrc.performance.simulation.PerformanceTestRunner
-import uk.gov.hmrc.perftests.digitaltariffs.DigitalTariffsPerformanceTestRunner
+import uk.gov.hmrc.perftests.digitaltariffs.Configuration
 import uk.gov.hmrc.perftests.digitaltariffs.traderui.AuthRequests._
 import uk.gov.hmrc.perftests.digitaltariffs.traderui.TraderUiRequests._
 
-class TraderUiSimulation extends PerformanceTestRunner with DigitalTariffsPerformanceTestRunner {
-
-  override val httpProtocol: HttpProtocolBuilder =
-    buildHttpProtocol(url = externalBaseUrl)
+class TraderUiSimulation extends PerformanceTestRunner with Configuration {
 
   setup("traderUI", "UK Trader applies for a BTI application") withRequests (
     // Government Gateway Sign In
